+++ doc/components/component/data/plugin/date.vue
@@ -22,6 +22,14 @@
     <div>
       <example demo="dataplugins/datepicker5"></example>
     </div>
+    <h3>选择周控件</h3>
+    <div>
+      <example demo="dataplugins/datepicker10"></example>
+    </div>
+    <!-- <h3>选择季度控件</h3>
+    <div>
+      <example demo="dataplugins/datepicker11"></example>
+    </div> -->
     <h3>快捷方式</h3>
     <p>可以通过全局定义配置好快捷方式，这里只做key的调用，或者自己定义函数返回。全局配置详情请参考<router-link to="/guide/config">全局配置</router-link></p>
     <div>
@@ -61,7 +69,7 @@
         <td>类型</td>
         <td>String</td>
         <td>date</td>
-        <td>year, month, week, datetime, datehour</td>
+        <td>year, month, week, datetime, datehour, week</td>
       </tr>
       <tr>
         <td>option</td>
@@ -97,6 +105,13 @@
         <td>String</td>
         <td>-</td>
         <td>请选择</td>
+      </tr>
+      <tr>
+        <td>startWeek</td>
+        <td>选择周的时候，设定星期的开始日</td>
+        <td>Number</td>
+        <td>1, 2, 3, 4, 5, 6, 7</td>
+        <td>1</td>
       </tr>
     </table>