+++ doc/components/demos/dataplugins/datepicker11.vue
@@ -1,15 +1,16 @@
 <template>
   <div>
       <p>value:{{value}}</p>
-      <DatePicker v-model="value" placeholder="请选择日期" type="month"></DatePicker>
+      <DatePicker v-model="value" placeholder="请选择日期" type="quarter"></DatePicker>
   </div>
 </template>
 
 <script>
+import manba from 'manba';
 export default {
   data() {
     return {
-      value: '',
+      value: ''
     }
   }
 }