+++ src/components/datetime/index.js
@@ -1,4 +1,5 @@
 import DatePicker from './datepicker';
 import DateRangePicker from './daterangepicker';
+import DateFullRangePicker from './datefullrangepicker';
 
-export default { DatePicker, DateRangePicker};
+export default { DatePicker, DateRangePicker, DateFullRangePicker };